+++ app/Forms/Builders/UserForm.php
@@ -23,6 +23,11 @@ class UserForm
 
     public function edit(User $user)
     {
+        if (auth()->user()->can('change-password', $user)) {
+            $user->password = null;
+            $this->form->show(['password', 'password_confirmation']);
+        }
+
         return $this->form->edit($user);
     }
 }