+++ resources/assets/js/pages/auth/password/Reset.vue
@@ -5,29 +5,9 @@
             route="password.reset"
             @success="success">
             <p slot="password-strength"
-                slot-scope="{ hasPassword, password }"
+                slot-scope="{ password }"
                 class="help">
-                <transition enter-active-class="zoomIn"
-                    leave-active-class="zoomOut">
-                    <svg class="animated"
-                        width="100%"
-                        height="5"
-                        version="1.1"
-                        xmlns="http://www.w3.org/2000/svg"
-                        v-if="hasPassword">
-                        <rect v-for="i in 5"
-                            width="15%"
-                            height="4"
-                            stroke-width="4"
-                            :x="2.5 + (i-1) * 15 + (i-1) * 5 + '%'"
-                            :key="i"
-                            :stroke="
-                                i <= score(hasPassword, password) + 1
-                                    ? 'green'
-                                    : 'orangered'
-                            "/>
-                    </svg>
-                </transition>
+                <password-strength :password="password"/>
             </p>
         </auth-form>
 
@@ -35,20 +15,15 @@
 
 <script>
 
-import zxcvbn from 'zxcvbn';
 import AuthForm from '../AuthForm.vue';
+import PasswordStrength from './PasswordStrength.vue';
 
 export default {
     name: 'Reset',
 
-    components: { AuthForm },
+    components: { AuthForm, PasswordStrength },
 
     methods: {
-        score(hasPassword, password) {
-            return hasPassword
-                ? zxcvbn(password).score
-                : 6;
-        },
         success({ status }) {
             this.$toastr.success(status);